pyproject.toml
[project]
name = "python-binance-trading-bot"
version = "0.1.0"
description = "Automated Binance trading bot with multiple strategies and backtesting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-binance>=1.0.19",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.10.2",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "psycopg2-binary>=2.9.0",
    "redis>=4.6.0",
    "schedule>=1.2.0",
    "plotly>=5.15.0",
    "dash>=2.12.0",
    "python-telegram-bot>=20.0",
    "yfinance>=0.2.18",
    "requests>=2.31.0",
    "websocket-client>=1.6.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
